+++ frontend/src/_test_/button.test.tsx
@@ -1,6 +1,7 @@
-import { render, fireEvent, screen } from '../setupTests'
-
+// import { fireEvent, screen } from '../setupTests'
+import { fireEvent, screen } from '@testing-library/react'
 import { ButtonGit } from '../components/atoms/button/index'
+import { render } from './config/test_utils'
 
 describe('Button', () => {
   const textBttn = 'Follow'
